+++ pelion_test_lib/cloud/libraries/rest_api/rest_api.py
@@ -1,5 +1,5 @@
 """
-Copyright 2019 ARM Limited
+Copyright 2019-2020 ARM Limited
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -18,9 +18,7 @@ import logging
 import requests
 from pelion_test_lib.tools.utils import assert_status
 
-logging.basicConfig(format="%(asctime)s:%(name)s:%(threadName)s:%(levelname)s: %(message)s")
 log = logging.getLogger(__name__)
-log.setLevel(logging.DEBUG)
 
 urllib3_logger = logging.getLogger('urllib3')
 urllib3_logger.setLevel(logging.WARNING)