+++ pelion_test_lib/tools/manifest_tool.py
@@ -1,5 +1,5 @@
 """
-Copyright 2019 ARM Limited
+Copyright 2019-2020 ARM Limited
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -20,9 +20,7 @@ import subprocess
 import os
 import uuid
 
-logging.basicConfig(format='%(asctime)s:%(name)s:%(threadName)s:%(levelname)s: %(message)s')
 log = logging.getLogger(__name__)
-log.setLevel(logging.DEBUG)
 
 BINARIES = 'bin'
 MANIFEST_TOOL = 'manifest-tool'